requirements.txt
yfinance==0.2.36
gspread==5.12.4
oauth2client==4.1.3
notion-client==2.2.1
pandas==2.1.4
numpy==1.26.3
streamlit==1.32.0
scikit-learn==1.4.0
requests==2.31.0
python-dotenv==1.0.1
pyyaml==6.0.1
plotly==5.18.0
gitpython==3.1.41 